requirements.txt
alembic==1.12.0
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2026.1.4
click==8.3.1
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.1.0
fastapi==0.104.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.1
idna==3.11
iniconfig==2.3.0
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.11.0
pytest==7.4.3
python-dotenv==1.2.1
python-multipart==0.0.6
PyYAML==6.0.3
redis==5.0.1
slowapi==0.1.8
sniffio==1.3.1
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.23.2
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.1
